backend_refactor init
