+++ .gitignore
@@ -34,5 +34,7 @@ debug.log
 Thumbs.db
 
 #db_cache
-my_folder/*
-!my_folder/.gitkeep
+db_cache/
+
+#chroma_db
+chroma_db/
