+++ backend/src/ai_model/requirements.txt
@@ -1,0 +1,7 @@
+google-cloud-storage
+langchain-google-genai
+langchain_community
+langchain-text-splitters
+langchain-chroma
+PyPDF2
+python-dotenv
