+++ backend/src/database/requirements.txt
@@ -1,0 +1,4 @@
+python-dotenv
+motor
+fastapi
+httpx
