+++ backend/src/routes/requirements.txt
@@ -1,0 +1,2 @@
+fastapi
+motor
