+++ scripts/stop_backend.sh
@@ -1,0 +1,1 @@
+kill -f uvicorn
